tobecontinued

+++ lib/utils.ts
@@ -25,7 +25,8 @@ export function getNow() {
 export const receiverBodySchema = z.object({
   image_url: z.string(),
   secret_key: z.string(),
-  confidence: z.number().float()
+  confidence: z.number() 
+
 });
 
 export type ReceiverBody = z.infer<typeof receiverBodySchema>;
